Shift headings in feeds
So basically, I have a `level` attribute on the `render_page` function that determines where in the document hierachy the thing gets rendered (if there are headings above it etc).

This change shifts the headings in the page itself to also conform to that setting. Aka if the title of the page is h3, any h2's in the content will become h4's.

I also created a nice API for capturing function output via the output buffering stuffz.